working feature edit & delete product + altered ddl for order_product table
Edit & delete product can only be done if product is not being process (on going job). Order table no longer uses productID, instead copy the value of product name and price from the product table, this will allow the product price and name in past transactions to not get affected if the product is changed. If a product is deleted then all corresponding product batches will be deleted